testsuite: Check in optional patch that offers to build everything against a custom glibc sysroot
This was developed because it seemed like glibc 2.33 was causing NSS-style functions that rely on
libgotcha's _dl_open() interception to fail.  However, the culprit appears to instead by global
variable interception.

Producing the necessary artifacts to support this requires additional libtestinger build steps (also
implemented in the patch) to avoid creating a shared object with too new a glibc version dependency.
One must also avoid dynamically linking against libstd because that too can have such a dependency;
unfortunately, Cargo makes this harder by *prepending* command-line rustc arguments to the rustflags
from the config TOML, so we have to use RUSTFLAGS instead as a workaround.  And all this while still
steering clear of the issue we worked around in 76ffca6.