handle.c: Fix longstanding performance regression affecting lazily-resolved symbol memoization
Our age-old "optimization" of not updating GOT entries for symbols we didn't interpose was
misimplemented for lazily-resolved symbols.  As promised, we did not update the GOT entry, but
tragically, we *did* tweak the relocation to trick the dynamic linker into memoizing symbol entries
into the SGOT instead of the GOT.  But because the SGOT wasn't actually on the lookup path, this
meant that lazy lookups were never actually memoized in this case and repeatedly took the slow path!

This was introduced all the way back in f923b82, and was finally discovered when, during a callgrind
profiling run, a signal handler was observed near the top of the cycle count ranking.  It turned out
it was spending most of its time in numerous calls to _dl_lookup_symbol_x(), most of which were
resolving __tls_get_addr() calls to access thread-local variables.